c/snap-bootstrap: split CVM related functionality in separate files
